+++ datawrangler/core/__init__.py
@@ -1,1 +1,7 @@
-from .configurator import get_default_options, apply_defaults, update_dict, __version__
+from .configurator import (get_default_options, apply_defaults, update_dict, __version__,
+                           set_dataframe_backend, get_dataframe_backend, reset_dataframe_backend)
+
+__all__ = [
+    'get_default_options', 'apply_defaults', 'update_dict', '__version__', 'set_dataframe_backend',
+    'get_dataframe_backend', 'reset_dataframe_backend'
+]