+++ datawrangler/decorate/__init__.py
@@ -1,1 +1,5 @@
 from .decorate import list_generalizer, funnel, interpolate, apply_stacked, apply_unstacked
+
+__all__ = [
+    'list_generalizer', 'funnel', 'interpolate', 'apply_stacked', 'apply_unstacked'
+]