+++ datawrangler/zoo/__init__.py
@@ -4,3 +4,9 @@ from .dataframe import is_dataframe, wrangle_dataframe, is_multiindex_dataframe
 from .null import is_null, wrangle_null
 from .text import is_text, wrangle_text, get_corpus, apply_text_model, get_text_model, to_str_list, get_text
 from ..util.helpers import dataframe_like, array_like
+
+__all__ = [
+    'wrangle', 'is_array', 'wrangle_array', 'is_dataframe', 'wrangle_dataframe', 'is_multiindex_dataframe',
+    'is_null', 'wrangle_null', 'is_text', 'wrangle_text', 'get_corpus', 'apply_text_model', 'get_text_model',
+    'to_str_list', 'get_text', 'dataframe_like', 'array_like'
+]